updates the codebase to MVC
