+++ routes/index.js
@@ -1,58 +1,24 @@
 // npm imports
 const express = require("express");
 
-// require models
-const Message = require("../models/message");
-const User = require("../models/user");
+// require controller
+const index_controller = require("../controllers/index-controller");
 
 // require routers
 const authRouter = require("./auth");
 const messagesRouter = require("./messages");
 
-// environment variables
-require("dotenv").config();
-
 const router = express.Router();
 
 /* GET home page. */
-router.get("/", function (req, res, next) {
-  Message.find({}, function (err, messages) {
-    if (err) {
-      return next(err);
-    }
-    res.render("index", { title: "ChatHouse", messages });
-  });
-});
+router.get("/", index_controller.index);
 
 router.use("/auth", authRouter);
 
 router.use("/messages", messagesRouter);
 
-router.get("/membership", function (req, res, next) {
-  res.render("membership", { title: "Membership" });
-});
+router.get("/membership", index_controller.membership_get);
 
-router.post("/membership", function (req, res, next) {
-  const { membership_passcode } = req.body;
-  const { _id } = req.user;
-
-  if (membership_passcode == process.env.MEMBERSHIP_PASSCODE) {
-    User.findByIdAndUpdate(
-      { _id },
-      { membership: true },
-      function (err, updated_user) {
-        if (err) {
-          return next(err);
-        }
-        res.redirect("/");
-      }
-    );
-  } else {
-    res.render("membership", {
-      title: "Membership",
-      errors: ["Incorrect Passcode"],
-    });
-  }
-});
+router.post("/membership", index_controller.membership_post);
 
 module.exports = router;
